Move reporters into their own packages
This will setup what will eventually be the plugins system
