+++ diff_cover/tests/test_violations_reporter.py
@@ -1,17 +1,21 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from mock import Mock, patch, MagicMock
+
 import subprocess
+import xml.etree.cElementTree as etree
 from subprocess import Popen
 from textwrap import dedent
-import xml.etree.cElementTree as etree
-from six import BytesIO, StringIO
+
 import six
-from diff_cover.violations_reporter import XmlCoverageReporter, Violation, \
-    Pep8QualityReporter, PyflakesQualityReporter, PylintQualityReporter, \
-    QualityReporterError, Flake8QualityReporter, JsHintQualityReporter, \
-    BaseQualityReporter
+from mock import Mock, patch, MagicMock
+from six import BytesIO, StringIO
+
 from diff_cover.tests.helpers import unittest
+from diff_cover.violationsreporters.base import QualityReporterError
+from diff_cover.violationsreporters.violations_reporter import XmlCoverageReporter, Violation, \
+    Pep8QualityReporter, PyflakesQualityReporter, PylintQualityReporter, \
+    Flake8QualityReporter, JsHintQualityReporter, \
+    BaseQualityReporter
 
 
 class XmlCoverageReporterTest(unittest.TestCase):
@@ -28,7 +32,7 @@ class XmlCoverageReporterTest(unittest.TestCase):
 
     def setUp(self):
         # Paths generated by git_path are always the given argument
-        _git_path_mock = patch('diff_cover.violations_reporter.GitPathTool').start()
+        _git_path_mock = patch('diff_cover.violationsreporters.violations_reporter.GitPathTool').start()
         _git_path_mock.relative_path = lambda path: path
         _git_path_mock.absolute_path = lambda path: path
 
@@ -1167,7 +1171,7 @@ class SubprocessErrorTestCase(unittest.TestCase):
         # on it, raise an OSError
         _mock_Popen = Mock()
         _mock_Popen.return_value.communicate.side_effect = OSError
-        patcher = patch("diff_cover.violations_reporter.subprocess.Popen", _mock_Popen)
+        patcher = patch("diff_cover.violationsreporters.base.subprocess.Popen", _mock_Popen)
         patcher.start()
         self.addCleanup(patcher.stop)
